InnerFunctionInterface
